錯誤修正完畢
git-svn-id: http://www.castiello-tsai.org:8523/svn/L1JTW@157 96aee97a-5c11-ed42-8518-d7f33e96f5f9

+++ src/l1j/server/server/model/Instance/L1PcInstance.java
@@ -1201,7 +1201,7 @@ public class L1PcInstance extends L1Character {
 				if (getLawful() < 0) {
 					lostRate *= 2;
 				}
-				short rnd = RandomArrayList.getArrayshortList(1000);
+				short rnd = RandomArrayList.getArrayshortList((short) 1000);
 				if (rnd <= lostRate) {
 					int count = 1;
 					if (getLawful() <= -30000) {
